fixed bug with button movement, added documentation for my preexisting/new code as I recall

+++ Physics/Physics/core/src/main/java/edu/cornell/cis3152/physics/level_player/enviromentals/Button.java
@@ -111,7 +111,7 @@ public class Button extends ObstacleGroup {
         );
 
         prismaticJointDef.enableLimit = true;
-        float adjustedButtonDepth = (float) (height);
+        float adjustedButtonDepth = height;
 
         prismaticJointDef.upperTranslation = 0;
         prismaticJointDef.lowerTranslation = -adjustedButtonDepth / (doubleSided ? 1 : 2);
